refactor: 사용자 인증 검증 추가, argumentResolver 패키지 변경 IllegalStateException 404 처리
